+++ tests/include/test_controllers/MemoryController.h
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2021-2024 Nicolas Beddows <nicolas.beddows@gmail.com>
+Copyright (c) 2021-2025 Nicolas Beddows <nicolas.beddows@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -113,7 +113,7 @@ namespace meen
 
 			@remark				This controller never generates any interrupts.
 		*/
-		ISR ServiceInterrupts(uint64_t currTime, uint64_t cycles, IController* controller) final;
+		ISR GenerateInterrupt(uint64_t currTime, uint64_t cycles, IController* controller) final;
 	};
 } // namespace meen