+++ CHANGELOG.md
@@ -57,6 +57,8 @@
 * The cpu will be reset when no cpu configutation has been
   specified by the `IMachine::OnLoad` handler.
 * The distribution documentation has changed from html to pdf.
+* Renamed `IController::ServiceInterrupts`
+  `IController::GenerateInterrupt`.
 
 1.6.2 [24/07/24]
 * Deprecated config options `ramOffset`, `ramSize`,